Extract stack-neutral Occurrent autoconfiguration into a shared module
Move OccurrentProperties, the Jackson3 CloudEventConverter configuration, the
CloudEventTypeMapper fallback, and the capability Condition classes out of the
blocking starter into a new spring-boot-autoconfigure-mongodb-common module.
These pieces do not depend on the blocking or reactive stack, so sharing them
lets a future reactive starter reuse the same configuration properties and
converter wiring without pulling in the blocking modules. A single shared
@ConfigurationProperties(prefix = "occurrent") class also means both starters
can sit on one classpath without a duplicate-bean clash.

The blocking starter now depends on the common module and imports the moved
classes from org.occurrent.springboot.mongo.common. Behavior is unchanged; the
full blocking test suite passes.

OccurrentProperties moved package (org.occurrent.springboot.mongo.blocking to
org.occurrent.springboot.mongo.common) and its javadoc now refers to the
blocking application-service and query types in prose instead of @link, since
the common module no longer depends on them.
