+++ framework/pom.xml
@@ -29,6 +29,7 @@
 
     <modules>
         <module>spring-boot-starter-mongodb</module>
+        <module>spring-boot-autoconfigure-mongodb-common</module>
         <module>annotations</module>
     </modules>
