+++ framework/spring-boot-autoconfigure-mongodb-common/src/main/java/org/occurrent/springboot/mongo/common/OnStreamEventStoreCapabilityCondition.java
@@ -15,7 +15,7 @@
  *  limitations under the License.
  */
 
-package org.occurrent.springboot.mongo.blocking;
+package org.occurrent.springboot.mongo.common;
 
 import org.jspecify.annotations.NonNull;
 import org.occurrent.eventstore.api.EventStoreCapability;
@@ -29,7 +29,7 @@ import java.util.Set;
 
 import static org.occurrent.eventstore.api.EventStoreCapability.STREAM;
 
-class OnStreamEventStoreCapabilityCondition implements Condition {
+public class OnStreamEventStoreCapabilityCondition implements Condition {
     @Override
     public boolean matches(ConditionContext context, @NonNull AnnotatedTypeMetadata metadata) {
         Set<EventStoreCapability> capabilities = Binder.get(context.getEnvironment())
