+++ framework/spring-boot-starter-mongodb/pom.xml
@@ -53,6 +53,11 @@
         <!-- Occurrent -->
         <dependency>
             <groupId>org.occurrent</groupId>
+            <artifactId>spring-boot-autoconfigure-mongodb-common</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.occurrent</groupId>
             <artifactId>eventstore-mongodb-spring-blocking</artifactId>
             <version>${project.version}</version>
         </dependency>
