+++ framework/spring-boot-starter-mongodb/src/main/java/org/occurrent/springboot/mongo/blocking/OccurrentMongoAutoConfiguration.java
@@ -43,7 +43,13 @@ import org.occurrent.eventstore.api.dcb.DcbQuery;
 import org.occurrent.eventstore.mongodb.spring.blocking.EventStoreConfig;
 import org.occurrent.eventstore.mongodb.spring.blocking.SpringMongoEventStore;
 import org.occurrent.retry.RetryStrategy;
-import org.occurrent.springboot.mongo.blocking.OccurrentProperties.EventStoreProperties;
+import org.occurrent.springboot.mongo.common.Jackson3CloudEventConverterConfiguration;
+import org.occurrent.springboot.mongo.common.OccurrentProperties;
+import org.occurrent.springboot.mongo.common.OccurrentProperties.EventStoreProperties;
+import org.occurrent.springboot.mongo.common.OnDcbEventStoreCapabilityCondition;
+import org.occurrent.springboot.mongo.common.OnDomainEventQueriesCapabilityCondition;
+import org.occurrent.springboot.mongo.common.OnMissingCloudEventConverterAndCloudEventTypeMapperCondition;
+import org.occurrent.springboot.mongo.common.OnStreamEventStoreCapabilityCondition;
 import org.occurrent.subscription.api.blocking.CompetingConsumerStrategy.CompetingConsumerListener;
 import org.occurrent.subscription.api.blocking.Subscribable;
 import org.occurrent.subscription.api.blocking.SubscriptionModel;
